+++ v3/crates/metadata-resolve/src/stages/model_permissions/model_permission.rs
@@ -373,6 +373,8 @@ pub(crate) fn resolve_model_predicate_with_type(
                     ValueExpression::SessionVariable(hasura_authn_core::SessionVariableReference {
                         name: session_variable.clone(),
                         passed_as_json: flags.contains(open_dds::flags::Flag::JsonSessionVariables),
+                        disallow_unknown_fields: flags
+                            .contains(open_dds::flags::Flag::DisallowUnknownValuesInArguments),
                     })
                 }
             };