+++ v3/crates/open-dds/src/flags.rs
@@ -44,6 +44,8 @@ pub enum Flag {
     DisallowUnsupportedOrderableRelationships,
     DisallowLocalRelationshipsOnDataConnectorsWithoutRelationshipsOrVariables,
     DisallowRecursiveObjectTypes,
+    DisallowUnknownValuesInArguments,
+    RequireValidCommandOutputType,
 }
 
 impl Flag {
@@ -119,6 +121,8 @@ impl Flag {
                 "disallow_local_relationships_on_data_connectors_without_relationships_or_variables"
             }
             Flag::DisallowRecursiveObjectTypes => "disallow_recursive_object_types",
+            Flag::DisallowUnknownValuesInArguments => "disallow_unknown_values_in_arguments",
+            Flag::RequireValidCommandOutputType => "require_valid_command_output_type",
         }
     }
 }